Create TPosts type
