+++ src/lib/hooks/useCreatePost.ts
@@ -10,25 +10,13 @@ import {
 } from '@tanstack/react-query';
 
 import { QueryKey } from '../enums';
-import {
-  type PostSchema,
-  type PostWithRelationsAndRelationCountsAndUserReaction,
-} from '../types';
+import type { PostSchema, TPosts } from '../types';
 
 import useSignedInUser from './useSignedInUser';
 
 type TPost = {
   data: { post: Post | null } | null;
   errors: { [key: string]: string[] } | unknown | null;
-  success: boolean;
-};
-
-type TPosts = {
-  data: {
-    nextCursor: number | null;
-    posts: PostWithRelationsAndRelationCountsAndUserReaction[];
-  };
-  errors: { [key: string]: string[] } | null;
   success: boolean;
 };
