+++ src/lib/hooks/useDeletePost.ts
@@ -10,20 +10,14 @@ import {
 } from '@tanstack/react-query';
 
 import { QueryKey } from '../enums';
-import { type PostWithRelationsAndRelationCountsAndUserReaction } from '../types';
+import type {
+  PostWithRelationsAndRelationCountsAndUserReaction,
+  TPosts,
+} from '../types';
 
 type TPost = {
   data: { post: Post | null } | null;
   errors: { [key: string]: string[] } | unknown | null;
-  success: boolean;
-};
-
-type TPosts = {
-  data: {
-    nextCursor: number | null;
-    posts: PostWithRelationsAndRelationCountsAndUserReaction[];
-  };
-  errors: { [key: string]: string[] } | null;
   success: boolean;
 };
