+++ src/lib/hooks/useDeleteReaction.ts
@@ -10,9 +10,10 @@ import {
 } from '@tanstack/react-query';
 
 import { QueryKey } from '../enums';
-import {
-  type CommentWithRelationsAndRelationCountsAndUserReaction,
-  type PostWithRelationsAndRelationCountsAndUserReaction,
+import type {
+  CommentWithRelationsAndRelationCountsAndUserReaction,
+  PostWithRelationsAndRelationCountsAndUserReaction,
+  TPosts,
 } from '../types';
 
 type TComments = {
@@ -27,15 +28,6 @@ type TComments = {
 type TPost = {
   data: { post: Post | null } | null;
   errors: { [key: string]: string[] } | unknown | null;
-  success: boolean;
-};
-
-type TPosts = {
-  data: {
-    nextCursor: number | null;
-    posts: PostWithRelationsAndRelationCountsAndUserReaction[];
-  };
-  errors: { [key: string]: string[] } | null;
   success: boolean;
 };
