+++ src/lib/hooks/useUpdatePost.ts
@@ -10,23 +10,15 @@ import {
 } from '@tanstack/react-query';
 
 import { QueryKey } from '../enums';
-import {
-  type PostSchema,
-  type PostWithRelationsAndRelationCountsAndUserReaction,
+import type {
+  PostSchema,
+  PostWithRelationsAndRelationCountsAndUserReaction,
+  TPosts,
 } from '../types';
 
 type TPost = {
   data: { post: Post | null } | null;
   errors: { [key: string]: string[] } | unknown | null;
-  success: boolean;
-};
-
-type TPosts = {
-  data: {
-    nextCursor: number | null;
-    posts: PostWithRelationsAndRelationCountsAndUserReaction[];
-  };
-  errors: { [key: string]: string[] } | null;
   success: boolean;
 };
