+++ src/lib/hooks/useUpdateReaction.ts
@@ -10,10 +10,11 @@ import {
 } from '@tanstack/react-query';
 
 import { QueryKey } from '../enums';
-import {
-  type CommentWithRelationsAndRelationCountsAndUserReaction,
-  type PostWithRelationsAndRelationCountsAndUserReaction,
-  type ReactionSchema,
+import type {
+  CommentWithRelationsAndRelationCountsAndUserReaction,
+  PostWithRelationsAndRelationCountsAndUserReaction,
+  ReactionSchema,
+  TPosts,
 } from '../types';
 
 type TComments = {
@@ -28,15 +29,6 @@ type TComments = {
 type TPost = {
   data: { post: Post | null } | null;
   errors: { [key: string]: string[] } | unknown | null;
-  success: boolean;
-};
-
-type TPosts = {
-  data: {
-    nextCursor: number | null;
-    posts: PostWithRelationsAndRelationCountsAndUserReaction[];
-  };
-  errors: { [key: string]: string[] } | null;
   success: boolean;
 };
